+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/storage/item/ItemStorageNBT.java
@@ -236,6 +236,14 @@ public abstract class ItemStorageNBT implements IItemStorage {
         return tag.getInteger(NBT_STORED);
     }
 
+    public static NBTTagCompound getNBTShareTag(NBTTagCompound tag) {
+        NBTTagCompound otherTag = new NBTTagCompound();
+
+        otherTag.setInteger(NBT_STORED, getStoredFromNBT(tag));
+
+        return otherTag;
+    }
+
     /*
      * @return A NBT tag initialized with the fields that {@link NBTStorage} uses
      */