+++ src/main/java/com/raoulvdberge/refinedstorage/item/ItemFluidStorageDisk.java
@@ -119,4 +119,9 @@ public class ItemFluidStorageDisk extends ItemBase {
     public int getEntityLifespan(ItemStack stack, World world) {
         return Integer.MAX_VALUE;
     }
+
+    @Override
+    public NBTTagCompound getNBTShareTag(ItemStack stack) {
+        return FluidStorageNBT.getNBTShareTag(stack.getTagCompound());
+    }
 }